# controllers/users.go
package controllers

import (
	"net/http"

	"github.com/carbans/netdebug/models"
	"github.com/carbans/netdebug/types"
	"github.com/carbans/netdebug/utils/token"
	"github.com/gin-gonic/gin"
)

type RegisterInput struct {
	Email    string `json:"email" binding:"required"`
	Password string `json:"password" binding:"required"`
}

type LoginInput struct {
	Email    string `json:"email" binding:"required"`
	Password string `json:"password" binding:"required"`
}

// Register registers a user
// @Summary Registers a user
// @Description Registers a user.
// @Accept json
// @Produce json
// @Tags user
// @Param register body RegisterInput true "Register Credentials"
// @Success 200
// @Failure 400 {object} types.ErrorResponse
// @Router /register [post]
func Register(c *gin.Context) {

	var input RegisterInput
	if err := c.ShouldBindJSON(&input); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	user := models.User{Email: input.Email, Password: input.Password}
	_, err := user.SaveUser()
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
	}

	c.JSON(http.StatusOK, gin.H{"data": "Correct register"})
}

// Login logs in a user
// @Summary Logs in a user
// @Description Login a user and returns authentication token.
// @Accept json
// @Produce json
// @Tags user
// @Param loginType body types.LoginType true "Login Credentials"
// @Success 200 {object} types.LoginResponseType "return token"
// @Failure 400 {object} types.ErrorResponse
// @Router /login [post]
func Login(c *gin.Context) {
	var input types.LoginType

	if err := c.ShouldBindJSON(&input); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	user := models.User{Email: input.Email, Password: input.Password}

	token, err := models.LoginCheck(user.Email, user.Password)

	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "username or password is incorrect."})
		return
	}

	c.JSON(http.StatusOK, gin.H{"token": token})
}

func CurrentUser(c *gin.Context) {
	user_id, err := token.ExtractTokenID(c)

	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	u, err := models.GetUserByID(user_id)

	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{"message": "success", "data": u})
}